splitted to separate files, added generic in mem storage
